refactor: export component
